+++ Applications/MessageQueues/Sender/main.c
@@ -9,12 +9,12 @@
 #include "../../ManualListEntry/apputils.h"
 
 void createQueueFile(void);
-void sendToQueue(const size_t* priorities, const size_t *priorityTypes, const size_t elementsCount);
+void sendToQueue(const size_t* priorities, const long* priorityTypes, const size_t elementsCount);
 
 int main()
 {
     const size_t priorities[] = {5, 2, 104, 205, 1200, 1099};
-    const size_t priorityTypes[] = {1, 1, 2, 2, 3, 3};
+    const long priorityTypes[] = {1, 1, 2, 2, 3, 3};
 
     createQueueFile();
 
@@ -53,7 +53,7 @@ void createQueueFile()
     }
 }
 
-void sendToQueue(const size_t* priorities, const size_t* priorityTypes, const size_t elementsCount)
+void sendToQueue(const size_t* priorities, const long* priorityTypes, const size_t elementsCount)
 {
     const key_t queueKey = ftok(QUEUE_FILENAME, BASE_ID);
     if (queueKey < 0)
